Lesson 15

+++ src/features/blogs/application/commands/remove-blog.handler.ts
@@ -2,11 +2,11 @@ import { BlogModel } from '../../entity/blog.schema';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { BlogsService } from '../blogs.service';
 import { BlogsRepositoryInterface } from '../../interfaces/blogs.repository.interface';
-import { Inject } from '@nestjs/common';
+import { ForbiddenException, Inject } from '@nestjs/common';
 import { BlogInjectionToken } from '../blog.injection.token';
 
 export class RemoveBlogCommand {
-	constructor(public id: string) {}
+	constructor(public id: string, public currentUserId: string) {}
 }
 
 @CommandHandler(RemoveBlogCommand)
@@ -19,6 +19,9 @@ export class RemoveBlogHandler implements ICommandHandler<RemoveBlogCommand> {
 
 	async execute(command: RemoveBlogCommand): Promise<void> {
 		const blog: BlogModel = await this.blogsService.findBlogOrErrorThrow(command.id);
+
+		if (blog.userId !== command.currentUserId) throw new ForbiddenException();
+
 		await this.blogsRepository.delete(blog);
 	}
 }
