+++ src/features/posts/application/commands/remove-post.handler.ts
@@ -2,11 +2,13 @@ import { PostModel } from '../../entity/post.schema';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { PostsService } from '../posts.service';
 import { PostsRepositoryInterface } from '../../interfaces/posts.repository.interface';
-import { Inject } from '@nestjs/common';
+import { ForbiddenException, Inject } from '@nestjs/common';
 import { PostInjectionToken } from '../post.injection.token';
+import { BlogModel } from '../../../blogs/entity/blog.schema';
+import { BlogsService } from '../../../blogs/application/blogs.service';
 
 export class RemovePostCommand {
-	constructor(public id: string) {}
+	constructor(public blogId: string, public postId: string, public currentUserId: string) {}
 }
 
 @CommandHandler(RemovePostCommand)
@@ -15,10 +17,14 @@ export class RemovePostHandler implements ICommandHandler<RemovePostCommand> {
 		@Inject(PostInjectionToken.POST_REPOSITORY)
 		private readonly postsRepository: PostsRepositoryInterface,
 		private readonly postsService: PostsService,
+		private readonly blogsService: BlogsService,
 	) {}
 
 	async execute(command: RemovePostCommand): Promise<void> {
-		const post: PostModel = await this.postsService.findPostOrErrorThrow(command.id);
+		const blog: BlogModel = await this.blogsService.findBlogOrErrorThrow(command.blogId);
+		if (blog.userId !== command.currentUserId) throw new ForbiddenException();
+
+		const post: PostModel = await this.postsService.findPostOrErrorThrow(command.postId);
 		await this.postsRepository.delete(post);
 	}
 }
